# Conflicts:
#	src/Sofco.ClientWeb/src/app/views/management-report/management-report.module.ts

+++ src/Sofco.ClientWeb/src/app/views/advancementAndRefund/refund/delegate/refund-delegate.ts
@@ -17,14 +17,22 @@ import { MenuService } from "app/services/admin/menu.service";
 export class RefundDelegateComponent implements OnInit, OnDestroy {
 
     public data: any[] = new Array<any>();
+    public dataApprovers: any[] = new Array<any>();
+    public dataDelegates: any[] = new Array<any>();
 
+    public managersAndDirectors: any[] = new Array<any>();
     public users: any[] = new Array<any>();
+
+    public allUsers: any[] = new Array<any>();
     public userId: number;
 
     public analytics: any[] = new Array();
     public analyticId: number;
 
     private itemSelected: any;
+
+    public type = "1";
+    analyticDisabled: boolean = false;
 
     public modalConfig: Ng2ModalConfig = new Ng2ModalConfig(
         'ACTIONS.confirmTitle',
@@ -49,6 +57,7 @@ export class RefundDelegateComponent implements OnInit, OnDestroy {
     private getAnalyticsSubscription: Subscription;
     private getUsersSubscription: Subscription;
     private getDataSubscription: Subscription;
+    private getData2Subscription: Subscription;
     private postSubscription: Subscription;
     private deleteSubscription: Subscription;
 
@@ -62,8 +71,9 @@ export class RefundDelegateComponent implements OnInit, OnDestroy {
 
     ngOnInit(): void {
         this.approversService.setType(UserApproverType.Refund);
-
+ 
         this.getUsers();
+        this.getManagersAndDirectors();
         this.getAnalytics();
         this.getData();
     }
@@ -74,6 +84,7 @@ export class RefundDelegateComponent implements OnInit, OnDestroy {
         if (this.getDataSubscription) this.getDataSubscription.unsubscribe();
         if (this.postSubscription) this.postSubscription.unsubscribe();
         if (this.deleteSubscription) this.deleteSubscription.unsubscribe();
+        if (this.getData2Subscription) this.getData2Subscription.unsubscribe();
     }
 
     getAnalytics() {
@@ -85,25 +96,90 @@ export class RefundDelegateComponent implements OnInit, OnDestroy {
     getUsers() {
         this.getUsersSubscription = this.userService.getOptions().subscribe(res => {
             this.users = res;
+            this.allUsers = res;
+        });
+    }
+
+    getManagersAndDirectors() {
+        this.getUsersSubscription = this.userService.getManagersAndDirectors().subscribe(res => {
+            this.managersAndDirectors = res;
         });
     }
 
     getData() {
+        var promises = new Array();
+
         var userId = this.menuService.user.id;
 
+        var promise1 = new Promise((resolve, reject) => {
+            this.getDataSubscription = this.approversService.getByUserId(userId).subscribe(res => {
+                this.dataApprovers = res.data.map(item => {
+                    item.checked = false;
+                    item.type = 'Aprobación';
+                    item.typeId = '1';
+                    return item;
+                });
+    
+                resolve();
+            },
+            error => resolve());
+        });
+
+        var promise2 = new Promise((resolve, reject) => {
+
+            this.getData2Subscription = this.refundService.getDelegates().subscribe(res => {
+                this.dataDelegates = res.data.map(delegate => {
+                    var item = {
+                        checked: false,
+                        id: delegate.id,
+                        approverName: delegate.text,
+                        analyticName: '',
+                        type: 'Generación',
+                        typeId: '2'
+                    }
+                    
+                    return item;
+                });
+    
+                resolve();
+            },
+            error => resolve());
+        });
+
+        promises.push(promise1);
+        promises.push(promise2);
+  
         this.messageService.showLoading();
 
-        this.getDataSubscription = this.approversService.getByUserId(userId).subscribe(res => {
+        Promise.all(promises).then(data => { 
             this.messageService.closeLoading();
 
-            this.data = res.data.map(item => {
-                item.checked = false;
-                return item;
+            this.data = [];
+
+            this.dataApprovers.forEach(x => {
+                this.data.push(x);
+            });
+
+            this.dataDelegates.forEach(x => {
+                this.data.push(x);
             });
 
             this.initTable();
-        },
-        error => this.messageService.closeLoading());
+        });
+    }
+
+    typeChange(){
+        this.userId = null;
+
+        if(this.type == '1'){
+            this.analyticDisabled = false;
+            this.allUsers = this.users;
+        }
+        else{
+            this.analyticDisabled = true;
+            this.analyticId = null;
+            this.allUsers = this.managersAndDirectors;
+        }
     }
 
     initTable() {
@@ -113,26 +189,45 @@ export class RefundDelegateComponent implements OnInit, OnDestroy {
     }
 
     save() {
-        var json ={
-            id: 0,
-            analyticId: this.analyticId,
-            ApproverUserId: this.userId,
-            userId: this.menuService.user.id,
-            employeeId: this.menuService.user.employeeId
+        if(this.type == '1'){
+            var json ={
+                id: 0,
+                analyticId: this.analyticId,
+                ApproverUserId: this.userId,
+                userId: this.menuService.user.id,
+                employeeId: this.menuService.user.employeeId
+            }
+
+            this.messageService.showLoading();
+
+            this.postSubscription = this.approversService.save([json]).subscribe(response => {
+                this.messageService.closeLoading();
+    
+                this.getData();
+            },
+            () => this.messageService.closeLoading());
         }
 
-        this.messageService.showLoading();
+        if(this.type == '2'){
+            this.messageService.showLoading();
 
-        this.postSubscription = this.approversService.save([json]).subscribe(response => {
-            this.messageService.closeLoading();
-
-            this.getData();
-        },
-        () => this.messageService.closeLoading());
+            this.postSubscription = this.refundService.addDelegate(this.userId).subscribe(response => {
+                this.messageService.closeLoading();
+    
+                this.getData();
+            },
+            () => this.messageService.closeLoading());
+        }
     }
 
     showSave(): boolean {
-        if(this.analyticId > 0 && this.userId > 0) return true;
+        if(this.type == '1' && this.analyticId > 0 && this.userId > 0){
+            return true;
+        }
+        
+        if(this.type == '2' && this.userId > 0){
+            return true;
+        }
 
         return false;
     }
@@ -144,18 +239,59 @@ export class RefundDelegateComponent implements OnInit, OnDestroy {
 
     processDelete() {
         const id = this.itemSelected.id;
-        this.confirmModal.hide();
 
-        this.deleteSubscription = this.approversService.delete(id).subscribe(response => {
-            this.getData();
-        }); 
+        if(this.itemSelected.typeId == '1'){
+            this.confirmModal.hide();
+
+            this.deleteSubscription = this.approversService.delete(id).subscribe(response => {
+                this.getData();
+            }); 
+        }
+
+        if(this.itemSelected.typeId == '2'){
+            this.confirmModal.hide();
+
+            this.deleteSubscription = this.refundService.deleteDelegate([id]).subscribe(response => {
+                this.getData();
+            }); 
+        }
     }
 
     processDeleteAll(){
-        const selecteds = this.data.filter(item => item.checked);
-        const ids = selecteds.map(x => x.id);
+        var promises = new Array();
 
-        this.deleteSubscription = this.approversService.deleteAll(ids).subscribe(response => {
+        var promise1 = new Promise((resolve, reject) => {
+            const approversSelecteds = this.data.filter(item => item.checked && item.typeId == '1');
+            const approversIds = approversSelecteds.map(x => x.id);
+    
+            if(approversIds && approversIds.length > 0){
+                this.deleteSubscription = this.approversService.deleteAll(approversIds).subscribe(response => {
+                    resolve();
+                });
+            }
+            else{
+                resolve();
+            }
+        });
+
+        var promise2 = new Promise((resolve, reject) => {
+            const delegatesSelecteds = this.data.filter(item => item.checked && item.typeId == '2');
+            const delegatesIds = delegatesSelecteds.map(x => x.id);
+    
+            if(delegatesIds && delegatesIds.length > 0){
+                this.deleteSubscription = this.refundService.deleteDelegate(delegatesIds).subscribe(response => {
+                    resolve();
+                });
+            }
+            else{
+                resolve();
+            }
+        });
+
+        promises.push(promise1);
+        promises.push(promise2);
+  
+        Promise.all(promises).then(data => { 
             this.confirmModal2.hide();
             this.getData();
         });
